+++ src/Feedy.Application/UseCases/GetPlaylistDetail/GetPlaylistDetailRequest.cs
@@ -1,0 +1,3 @@
+namespace Feedy.Application.UseCases.GetPlaylistDetail;
+
+public record GetPlaylistDetailRequest(Guid PlaylistId, Guid UserId);