+++ src/Feedy.Application/UseCases/GetSubscribedPlaylists/GetSubscribedPlaylistsResponse.cs
@@ -1,0 +1,3 @@
+namespace Feedy.Application.UseCases.GetSubscribedPlaylists;
+
+public record GetSubscribedPlaylistsResponse(IEnumerable<PlaylistSummaryDto> Items);